relocate some files
