--- datasource/remote/search/src/main/java/com/example/search/GenresRemoteDataSource.kt
@@ -1,7 +1,0 @@
-package com.example.search
-
-import com.example.search.models.GenresDto
-
-fun interface GenresRemoteDataSource {
-    suspend fun getAllGenres(): GenresDto
-}
