+++ datasource/remote/search/src/main/java/com/example/search/GenresRemoteDataSourceImp.kt
@@ -1,9 +1,11 @@
 package com.example.search
 
 import com.example.search.service.contract.GenresApiServices
-import com.example.search.models.GenresDto
+import com.repository.search.dataSource.remote.GenresRemoteDataSource
+import com.repository.search.dto.GenresDto
 
-class GenresRemoteDataSourceImp(private val apiService: GenresApiServices) : GenresRemoteDataSource {
+class GenresRemoteDataSourceImp(private val apiService: GenresApiServices) :
+    GenresRemoteDataSource {
     override suspend fun getAllGenres(): GenresDto {
         return apiService.getAllGenres()
     }
