+++ datasource/remote/search/src/main/java/com/example/search/SearchRemoteDataSourceImpl.kt
@@ -1,9 +1,11 @@
 package com.example.search
 
-import com.example.search.models.SearchDto
 import com.example.search.service.contract.SearchApiService
+import com.repository.search.dataSource.remote.SearchRemoteDataSource
+import com.repository.search.dto.SearchDto
 
-class SearchRemoteDataSourceImpl(private val apiService: SearchApiService) : SearchRemoteDataSource {
+class SearchRemoteDataSourceImpl(private val apiService: SearchApiService) :
+    SearchRemoteDataSource {
 
     override suspend fun searchMulti(query: String, page: Int, language: String): SearchDto {
         return apiService.searchMulti(query, page, language)
