+++ datasource/remote/search/src/main/java/com/example/search/di/SearchModule.kt
@@ -1,13 +1,11 @@
 package com.example.search.di
 
-import com.repository.search.dataSource.remote.GenresRemoteDataSource
 import com.example.search.GenresRemoteDataSourceImp
-import com.repository.search.dataSource.remote.SearchRemoteDataSource
 import com.example.search.SearchRemoteDataSourceImpl
 import com.example.search.service.contract.GenresApiServices
+import com.example.search.service.contract.SearchApiService
 import com.example.search.service.implementation.KtorGenresApiServices
 import com.example.search.service.implementation.KtorSearchApiService
-import com.example.search.service.contract.SearchApiService
 import org.koin.core.module.Module
 import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
