+++ datasource/remote/search/src/main/java/com/example/search/service/contract/GenresApiServices.kt
@@ -1,6 +1,6 @@
 package com.example.search.service.contract
 
-import com.example.search.models.GenresDto
+import com.repository.search.dto.GenresDto
 
 fun interface GenresApiServices {
     suspend fun getAllGenres(): GenresDto
