+++ datasource/remote/search/src/main/java/com/example/search/service/contract/SearchApiService.kt
@@ -1,14 +1,15 @@
 package com.example.search.service.contract
 
-import com.repository.search.models.SearchDto
+import com.repository.search.dto.SearchDto
+
 
 interface SearchApiService {
-    suspend fun searchMulti(query: String, page: Int, language: String): com.repository.search.models.SearchDto
-    suspend fun searchPerson(query: String, page: Int, language: String): com.repository.search.models.SearchDto
+    suspend fun searchMulti(query: String, page: Int, language: String): SearchDto
+    suspend fun searchPerson(query: String, page: Int, language: String): SearchDto
     suspend fun searchCountryCode(
         query: String,
         page: Int,
         language: String,
         countryCode: String,
-    ): com.repository.search.models.SearchDto
+    ): SearchDto
 }
