+++ datasource/remote/search/src/main/java/com/example/search/service/implementation/KtorGenresApiServices.kt
@@ -1,7 +1,7 @@
 package com.example.search.service.implementation
 
-import com.example.search.models.GenresDto
 import com.example.search.service.contract.GenresApiServices
+import com.repository.search.dto.GenresDto
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
