+++ datasource/remote/search/src/main/java/com/example/search/service/implementation/KtorSearchApiService.kt
@@ -1,10 +1,11 @@
 package com.example.search.service.implementation
 
-import com.example.search.models.SearchDto
 import com.example.search.service.contract.SearchApiService
-import io.ktor.client.*
-import io.ktor.client.call.*
-import io.ktor.client.request.*
+import com.repository.search.dto.SearchDto
+import io.ktor.client.HttpClient
+import io.ktor.client.call.body
+import io.ktor.client.request.get
+import io.ktor.client.request.parameter
 
 class KtorSearchApiService(
     private val httpClient: HttpClient,
