+++ repository/search/src/main/java/com/repository/search/dto/GenresDto.kt
@@ -1,4 +1,4 @@
-package com.example.search.models
+package com.repository.search.dto
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
